+++ src/server/chapter.rs
@@ -16,11 +16,11 @@ cfg_if! {
 
 #[server(AddChapter, "/secure", "Url", "add_chapter")]
 pub async fn add_chapter(
-    book_id: i64,
+    book_id: i32,
     title: String,
     closing_time: String,
     events: Vec<EventContent>,
-) -> Result<i64, ServerFnError> {
+) -> Result<i32, ServerFnError> {
     let book_sub = get_book(book_id).await?;
 
     match book_sub.role {
@@ -77,7 +77,7 @@ pub async fn add_chapter(
 }
 
 #[server(GetChapters, "/secure", "Url", "get_chapters")]
-pub async fn get_chapters(book_id: i64) -> Result<Vec<Chapter>, ServerFnError> {
+pub async fn get_chapters(book_id: i32) -> Result<Vec<Chapter>, ServerFnError> {
     let book_subscription = get_book(book_id).await?;
     match book_subscription.role {
         BookRole::Unauthorized => {
@@ -105,7 +105,7 @@ pub async fn get_chapters(book_id: i64) -> Result<Vec<Chapter>, ServerFnError> {
 }
 
 #[server(GetChapter, "/secure", "Url", "get_chapter")]
-pub async fn get_chapter(chapter_id: i64) -> Result<Chapter, ServerFnError> {
+pub async fn get_chapter(chapter_id: i32) -> Result<Chapter, ServerFnError> {
     let pool = pool()?;
 
     let book_id = sqlx::query!(
@@ -153,7 +153,7 @@ pub async fn get_chapter(chapter_id: i64) -> Result<Chapter, ServerFnError> {
 }
 
 #[server(SetOpen, "/secure", "Url", "set_open")]
-pub async fn set_open(chapter_id: i64, new_status: bool) -> Result<(), ServerFnError> {
+pub async fn set_open(chapter_id: i32, new_status: bool) -> Result<(), ServerFnError> {
     let pool = pool()?;
 
     let book_id = sqlx::query!(
@@ -193,7 +193,7 @@ pub async fn set_open(chapter_id: i64, new_status: bool) -> Result<(), ServerFnE
 }
 
 #[server(IsOpen, "/secure", "Url", "is_open")]
-pub async fn is_open(chapter_id: i64) -> Result<bool, ServerFnError> {
+pub async fn is_open(chapter_id: i32) -> Result<bool, ServerFnError> {
     let pool = pool()?;
 
     let book_id = sqlx::query!(
@@ -231,7 +231,7 @@ pub async fn is_open(chapter_id: i64) -> Result<bool, ServerFnError> {
 }
 
 #[server(GetEvents, "/secure", "Url", "get_events")]
-pub async fn get_events(chapter_id: i64) -> Result<Vec<Event>, ServerFnError> {
+pub async fn get_events(chapter_id: i32) -> Result<Vec<Event>, ServerFnError> {
     let pool = pool()?;
 
     let book_id = sqlx::query!(
@@ -273,7 +273,7 @@ pub async fn get_events(chapter_id: i64) -> Result<Vec<Event>, ServerFnError> {
 }
 
 #[server(GetPick, "/secure", "Url", "get_pick")]
-pub async fn get_pick(event_id: i64) -> Result<Pick, ServerFnError> {
+pub async fn get_pick(event_id: i32) -> Result<Pick, ServerFnError> {
     let user = auth()?.current_user.unwrap();
     let pool = pool()?;
 
@@ -329,7 +329,7 @@ pub async fn get_pick(event_id: i64) -> Result<Pick, ServerFnError> {
 
 #[server(GetPicks, "/secure", "Url", "get_picks")]
 pub async fn get_picks(
-    chapter_id: i64,
+    chapter_id: i32,
 ) -> Result<Vec<(String, Vec<(Event, Pick)>)>, ServerFnError> {
     let events = get_events(chapter_id)
         .await
@@ -436,7 +436,7 @@ pub async fn save_picks(picks: Vec<Pick>) -> Result<(), ServerFnError> {
 }
 
 #[server(GetUserInputs, "/secure", "Url", "get_user_inputs")]
-pub async fn get_user_inputs(event_id: i64) -> Result<Vec<String>, ServerFnError> {
+pub async fn get_user_inputs(event_id: i32) -> Result<Vec<String>, ServerFnError> {
     let pool = pool()?;
 
     let chapter_id = sqlx::query!(
@@ -471,7 +471,7 @@ pub async fn get_user_inputs(event_id: i64) -> Result<Vec<String>, ServerFnError
 }
 
 #[server(SaveAnswers, "/secure", "Url", "save_answers")]
-pub async fn save_answers(picks: Vec<(i64, Vec<String>)>) -> Result<(), ServerFnError> {
+pub async fn save_answers(picks: Vec<(i32, Vec<String>)>) -> Result<(), ServerFnError> {
     let pool = pool()?;
 
     let book_id = sqlx::query!(
@@ -525,13 +525,13 @@ pub async fn save_answers(picks: Vec<(i64, Vec<String>)>) -> Result<(), ServerFn
 }
 
 // #[server(GetChapterTable, "/secure", "Url", "get_chapter_table")]
-// pub async fn get_chapter_table(chapter_id: i64) -> Result<String, ServerFnError> {
+// pub async fn get_chapter_table(chapter_id: i32) -> Result<String, ServerFnError> {
 // 	if is_open(chapter_id).await? {
 // 		return Err(ServerFnError::Request("The chapter isn't closed yet! You can't see everyone's picks!".into()))
 // 	}
 
 // 	let events = get_events(chapter_id).await?;
-// 	let mut teams: HashMap<i64, Team> = HashMap::new();
+// 	let mut teams: HashMap<i32, Team> = HashMap::new();
 
 // 	for event in events.iter() {
 // 		match &event.contents {
